--- src/dataBundles/UniversalHourBlock.java
@@ -1,11 +1,0 @@
-package dataBundles;
-
-import java.time.DayOfWeek;
-import java.time.LocalTime;
-
-public interface UniversalHourBlock {
-    LocalTime startTime();
-    LocalTime endTime();
-    DayOfWeek dayOfWeek();
-
-}
