+++ src/dataBundles/UniversalTimeBlockWithDay.java
@@ -1,0 +1,13 @@
+package dataBundles;
+
+import java.time.DayOfWeek;
+import java.time.LocalDate;
+
+public interface UniversalTimeBlockWithDay extends UniversalTimeBlock {
+    LocalDate day();
+
+    @Override
+    default DayOfWeek dayOfWeek(){
+        return day().getDayOfWeek();
+    }
+}
